web_app/requirements.txt
albumentations==1.4.18
numpy==1.24.4
onnx==1.17.0
onnxscript==0.1.0
pillow==10.4.0
scikit-learn==1.3.2
streamlit==1.36.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.4
